+++ crates/biome_configuration/src/generated.rs
@@ -3,7 +3,7 @@
 use crate::linter::*;
 use crate::{RuleConfiguration, Rules};
 use biome_analyze::{AnalyzerRules, MetadataRegistry};
-pub(crate) fn push_to_analyzer_rules(
+pub fn push_to_analyzer_rules(
     rules: &Rules,
     metadata: &MetadataRegistry,
     analyzer_rules: &mut AnalyzerRules,
