+++ crates/biome_configuration/src/linter/mod.rs
@@ -2,21 +2,17 @@
 mod rules;
 
 pub use crate::linter::rules::Rules;
-use crate::overrides::OverrideLinterConfiguration;
 use biome_deserialize::{
     DeserializableValue, DeserializationDiagnostic, Merge, StringSet, VisitableType,
 };
 use biome_deserialize_macros::{Deserializable, Merge, Partial};
 use biome_diagnostics::Severity;
 use biome_js_analyze::options::PossibleOptions;
-use biome_service::settings::{to_matcher, LinterSettings};
-use biome_service::{Matcher, WorkspaceError};
 use bpaf::Bpaf;
 pub use rules::*;
 #[cfg(feature = "schema")]
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-use std::path::PathBuf;
 use std::str::FromStr;
 
 #[derive(Clone, Debug, Deserialize, Eq, Partial, PartialEq, Serialize)]
@@ -63,31 +59,6 @@ impl Default for LinterConfiguration {
 impl PartialLinterConfiguration {
     pub const fn is_disabled(&self) -> bool {
         matches!(self.enabled, Some(false))
-    }
-}
-
-pub fn to_linter_settings(
-    working_directory: Option<PathBuf>,
-    conf: LinterConfiguration,
-) -> Result<LinterSettings, WorkspaceError> {
-    Ok(LinterSettings {
-        enabled: conf.enabled,
-        rules: Some(conf.rules),
-        ignored_files: to_matcher(working_directory.clone(), Some(&conf.ignore))?,
-        included_files: to_matcher(working_directory.clone(), Some(&conf.include))?,
-    })
-}
-
-impl TryFrom<OverrideLinterConfiguration> for LinterSettings {
-    type Error = WorkspaceError;
-
-    fn try_from(conf: OverrideLinterConfiguration) -> Result<Self, Self::Error> {
-        Ok(Self {
-            enabled: conf.enabled.unwrap_or_default(),
-            rules: conf.rules,
-            ignored_files: Matcher::empty(),
-            included_files: Matcher::empty(),
-        })
     }
 }
