+++ crates/biome_service/src/file_handlers/json.rs
@@ -1,4 +1,5 @@
 use super::{ExtensionHandler, Mime};
+use crate::configuration::to_analyzer_rules;
 use crate::file_handlers::{
     AnalyzerCapabilities, Capabilities, FixAllParams, FormatterCapabilities, LintParams,
     LintResults, ParserCapabilities,
@@ -13,7 +14,7 @@ use crate::workspace::{
 };
 use crate::WorkspaceError;
 use biome_analyze::{AnalyzerConfiguration, AnalyzerOptions, ControlFlow, Never, RuleCategories};
-use biome_configuration::{to_analyzer_rules, PartialConfiguration, Rules};
+use biome_configuration::{PartialConfiguration, Rules};
 use biome_deserialize::json::deserialize_from_json_ast;
 use biome_diagnostics::{category, Diagnostic, DiagnosticExt, Severity};
 use biome_formatter::{FormatError, IndentStyle, IndentWidth, LineEnding, LineWidth, Printed};
